[ARV-32] refactor: MemberCommandTest로 병합 및 주석 추가
