+++ src/test/java/com/backend/allreva/member/command/MemberCommandTest.java
@@ -3,23 +3,29 @@ package com.backend.allreva.member.command;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.backend.allreva.member.command.application.MemberCommandService;
+import com.backend.allreva.member.command.application.MemberRepository;
 import com.backend.allreva.member.command.application.dto.MemberInfoUpdateRequest;
 import com.backend.allreva.member.command.domain.Member;
 import com.backend.allreva.member.command.domain.value.LoginProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.util.ReflectionTestUtils;
 
+@ExtendWith(MockitoExtension.class)
 @SuppressWarnings("NonAsciiCharacters")
-public class MemberUpdateCommandTest {
+public class MemberCommandTest {
 
-    private final MemberCommandService memberCommandService;
+    @InjectMocks
+    private MemberCommandService memberCommandService;
+
+    @Mock
+    private MemberRepository memberRepository;
 
     Member member;
-
-    public MemberUpdateCommandTest() {
-        this.memberCommandService = new MemberCommandService();
-    }
 
     @BeforeEach
     void setUp() {
@@ -29,7 +35,22 @@ public class MemberUpdateCommandTest {
                 LoginProvider.GOOGLE,
                 "https://my_picture");
         ReflectionTestUtils.setField(member, "id", 1L);
-        member.upgradeToUser();
+    }
+
+    @Test
+    void 회원_가입_시_회원_정보를_성공적으로_입력한다() {
+        // given
+        MemberInfoUpdateRequest memberInfoUpdateRequest = new MemberInfoUpdateRequest(
+                "updated nickname",
+                "test introduce",
+                "updated profile image url"
+        );
+
+        // when
+        Member updatedMember = memberCommandService.updateMemberInfo(memberInfoUpdateRequest, member);
+
+        // then
+        assertThat(updatedMember.getMemberInfo().getIntroduce()).isEqualTo("test introduce");
     }
 
     @Test
